backend/requirements.txt
flask>=3.0
flask-cors>=4.0.0
Flask-SocketIO>=5.4.1
eventlet>=0.36.1
pymongo>=4.8.0
pydantic>=2.6.0
bcrypt>=4.1.3
PyJWT>=2.9.0
celery>=5.4.0
redis>=5.0.4
razorpay>=1.4.2
scikit-learn>=1.5.1
joblib>=1.4.2
Flask-Limiter>=3.8.0
python-json-logger>=2.0.7
gunicorn>=22.0.0
python-dotenv==1.0.1
passlib==1.7.4
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0


flask-pymongo
werkzeug
pyjwt
